backend/requirements.txt

fastapi>=0.115.0
uvicorn>=0.30.6
sqlalchemy>=2.0.35
psycopg2-binary>=2.9.9
python-dotenv>=1.0.1
pydantic>=2.0
pydantic-settings
langchain>=0.2.0
langchain-anthropic>=0.1.0
langgraph>=0.1.0
langchain-core>=0.2.0
langchain-community>=0.2.0

bcrypt==4.0.1
python-jose[cryptography]==3.5.0
redis==5.0.1
rapidfuzz==3.6.1

